+++ config/openshift/v4_18/validate.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.)
 
-package v4_18_exp
+package v4_18
 
 import (
 	"github.com/coreos/butane/config/common"